Split out hir-def attribute handling parts into hir-expand
